Move 2FA classes into totp subpackage
