+++ backend/src/main/java/com/swappable/backend/auth/AuthController.java
@@ -16,6 +16,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import java.util.Map;
+import com.swappable.backend.auth.totp.TwoFactorService;
 
 @RestController
 @RequestMapping("/api/auth")
