+++ backend/src/main/java/com/swappable/backend/auth/totp/TwoFactorController.java
@@ -1,4 +1,4 @@
-package com.swappable.backend.auth;
+package com.swappable.backend.auth.totp;
 
 import com.swappable.backend.user.User;
 import com.swappable.backend.user.UserRepository;
@@ -7,6 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
+import com.swappable.backend.auth.totp.TwoFactorService;
 
 import java.util.Map;
